Make code generator tests usable from other slick-extensions:
slick-testkit now depends on slick-codegen and provides the
TestCodeGenerator and TestCodeRunner types which provide the tools
needed to generate and run code generator tests.
