+++ slick-testkit/src/test/scala/slick/test/codegen/GeneratedCodeTest.scala
@@ -1,44 +1,16 @@
 package slick.test.codegen
 
-import java.util.concurrent.ExecutionException
-
-import org.junit.Test
-import org.junit.Assert._
-import scala.concurrent.Await
 import scala.concurrent.ExecutionContext.Implicits.global
-import scala.concurrent.duration.Duration
 import slick.jdbc.meta.MTable
 import slick.ast.{Node, Select}
 import slick.test.codegen.generated._
-import com.typesafe.slick.testkit.util.JdbcTestDB
+import com.typesafe.slick.testkit.util.{TestCodeRunner, JdbcTestDB}
+import org.junit.Assert._
 
 /** Test files generated by CodeGeneratorTest */
-class GeneratedCodeTest {
-  def run(cln: String): Unit = {
-    val t = Class.forName(cln+"$").getField("MODULE$").get(null).asInstanceOf[GeneratedCodeTest.TestCase]
-    val tdb = t.tdb
-    println(s"Running test $cln on ${tdb.confName}")
-    if(tdb.isEnabled) {
-      tdb.cleanUpBefore()
-      try {
-        val a = t.test
-        val db = tdb.createDB()
-        try Await.result(db.run(a.withPinnedSession), Duration.Inf)
-        catch { case e: ExecutionException => throw e.getCause }
-        finally db.close()
-      } finally tdb.cleanUpAfter()
-    } else println("- Test database is disabled")
-  }
-
-  @Test def allTests = AllTests.clns.foreach(run)
-}
+class GeneratedCodeTest extends TestCodeRunner(AllTests)
 
 object GeneratedCodeTest {
-  trait TestCase {
-    def test: slick.dbio.DBIO[Any]
-    def tdb: JdbcTestDB
-  }
-
   def testCG1 = {
     import CG1._
     import tdb.driver.api._
